% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlpsb.R
\docType{package}
\name{anlpsbff}
\alias{anlpsbff}
\alias{anlpsbff-package}
\title{ANLP-SB}
\description{
Bayesian sparse multivariate regression model with variable selection using asymetric nonlocal priors (NLPs)
}
\details{
ANLP-SB models are fit using the \code{\link{anlpsbm}} function
}
\references{
\emph{A Bayesian Sparse Multivariate Regression Model with Asymmetric Nonlocal Priors for Microbiome Data}
}
\author{
Kurtis Shuler
}
