% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlpsb.R
\name{anlpsbm}
\alias{anlpsbm}
\title{ANLP-SB Model}
\usage{
anlpsbm(Y.mat, X.mat, time, rep.K, Miss.var.ind.mat = NULL,
  max.miss.cat.vec = NULL, first.p.missing.idx1.vec = NULL, PRIOR,
  SETTINGS, fftempdir)
}
\arguments{
\item{Y.mat}{\eqn{N x J} matrix of counts}

\item{X.mat}{\eqn{n x P} design matrix}

\item{time}{\eqn{n}-dimensional vector of observation timepoints}

\item{rep.K}{\eqn{n}-dimensional vector of the number of replicates at each timepoint}

\item{Miss.var.ind.mat}{\eqn{n x Q} (0, 1) matrix where \eqn{Q} is the number of covariates with missing values.  A value of "1" in the matrix indicates the covariate value is missing at that timepoint, and a "0" indicates that covariate value was observed.}

\item{max.miss.cat.vec}{\eqn{Q}-dimensional vector where the \eqn{q}th element has the number of factor levels in X.mat for the missing covariates.}

\item{first.p.missing.idx1.vec}{\eqn{Q}-dimensional vector with the \eqn{q}th element indicating the column index where the missing covariate occurs in X.mat}

\item{PRIOR}{List of the model hyperparameters with the following elements:
\itemize{
    \item a.h; normal mean for h
    \item b.h.2; normal variance for h
    \item a.kappa.2; normal mean for kappa.2
    \item b.kappa.2; normal variance for kappa.2
    \item a.psi.r; dirichlet concentration parameter for psi.r
    \item a.w.r; beta shape parameter for w.r
    \item b.w.r; beta shape parameter for w.r
    \item u.r.w; normal variance for r.tk mixture
    \item upsilon.r; r.tk mean constraint
    \item b.eta.r.2; normal variance for eta.r
    \item a.psi.alpha0; Dirichlet concentration parameter for psi.alpha0
    \item a.w.alpha0; beta shape parameter for w.alpha0
    \item b.w.alpha0; beta shape parameter for w.alpha0
    \item u.alpha0.2; normal variance for alpha.0j mixture
    \item upsilon.alpha0; alpha.0j mean constraint
    \item b.eta.alpha0.2; normal variance for eta.alpha
    \item a.tau; inverse-gamma shape paramter for tau.2
    \item b.tau; inverse-gamma scale parameter for tau.2
    \item a.pi.0; beta shape parameter for pi.0
    \item b.pi.0; beta shape parameter for pi.0
    \item a.pi.1; beta shape paramter for pi.1
    \item b.pi.1; beta shape paramter for pi.1
    \item a.sigma; inverse-gamma shape parameter for sigma.2.p
    \item b.sigma; inverse-gamma scale parameter for sigma.2.p
    \item a.iota; gamma shape parameter for iota.p
    \item b.iota; gamma rate parameter for iota.p
}
    The companion paper in the references has more details on the prior specification.}

\item{SETTINGS}{List of model settings with the following elements:
\itemize{
    \item n.mcmc; number of MCMC iterations
    \item print.mcmc; print MCMC progress every print.mcmc iterations.  Leave NULL to supress progress output.
    \item n.thin; MCMC thinning interval
    \item u.m.basis; knot points for process convolutions
    \item phi; variance/range parameter for process convolutions kernel
    \item s.j.tilde.proposal.sd; SD for Metropolis proposal of s.j
    \item L.r.trunc; number of mixture components for r.tk prior
    \item w.ell.r.proposal.sd; SD for Metropolis proposal of w.ell.r
    \item r.tk.tilde.proposal.sd; SD for Metropolis proposal of r.tk
    \item L.alpha0.trunc; number of mixture components for alpha0j prior
    \item w.ell.alpha0.proposal.sd; SD for Metropolis proposal of w.ell.alpha
    \item alpha0.j.proposal.sd; SD for Metropolis proposal of alpha.0.j
    \item theta.mj.proposal.sd; SD for Metropolis proposal of theta.mj
    \item beta.jp.proposal.sd; SD for Metropolis proposal of beta.jp
    \item beta.jp.joint.proposal.sd; step-size for joint Metropolis proposal of beta.jp for all P
    \item beta.jp.joint.otu.proposal.sd; step-size for joint Metropolis proposal of beta.jp for all J
    \item iota.p.proposal.sd; SD for Metropolis proposal of iota.p
    \item iota.p.joint.proposal.sd; step-size for joint Metropolis proposal of iota.p
    \item sigma.p.2.joint.proposal.sd; step-size for joint Metropolis proposal of sigma.p.2
}}
}
\value{
List with elements:
\itemize{
    \item samps; MCMC samples
    \item prior; prior specification
    \item settings; model settings
}
}
\description{
Fits ANLP-SB model on count data.  Details of the model description are available in the companion paper listed in the references.
}
\references{
\emph{A Bayesian Sparse Multivariate Regression Model with Asymmetric Nonlocal Priors for Microbiome Data}
}
