% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{fnBetaGammaIotaConsistent}
\alias{fnBetaGammaIotaConsistent}
\title{This function is only for testing purposes.  Checks if iota.vec is consistent
with Beta.mat}
\usage{
fnBetaGammaIotaConsistent(Beta.mat, Gamma.mat, iota.vec, sigma.2.vec)
}
\description{
This function is only for testing purposes.  Checks if iota.vec is consistent
with Beta.mat
}
